+++ processor/src/test/fixtures/expected/SingletonWithIdModelDaggerModule.java
@@ -3,6 +3,7 @@ import dagger.Module;
 import javax.annotation.Generated;
 import javax.inject.Singleton;
 
+@SuppressWarnings("DefaultAnnotationParam")
 @Generated("arez.processor.ArezProcessor")
 @Module
 public interface SingletonWithIdModelDaggerModule {