+++ processor/src/test/fixtures/input/com/example/component/DeprecatedModel.java
@@ -1,0 +1,9 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+
+@Deprecated
+@ArezComponent( allowEmpty = true )
+public abstract class DeprecatedModel
+{
+}